added a Ex3.c file

+++ Ex3.c
@@ -1,6 +1,6 @@
 #include<stdio.h>
 int main(){
-    unsigned long long int a=1;
+    unsigned long int a=1;
     int count=0;
     while(a!=0){
         a=a<<1;
